Started Implementing First Heuristic
Changed some things about how we represent points and nodes. Added a
description of the parts of the algorithm, and added the code for move
and search. Need to add the actual A* part, and write some version of
polling. The polling part will be easy to change though to rapidly try
different things.